Unit Converter
The Unit Converter is a straightforward, easy-to-use desktop conversion tool that allows you to convert between multiple length measurements quickly. Whether it's meters, feet, inches, miles or other common units, all you need to do is type a value, select the unit you want to convert from, and select the unit you want to convert to, and behold as the calculator will give you an answer in seconds. The unit converter was meant to be simple and always running when you need it to be, without the hassle of needing to use a calculator or input complicated settings.